Reorganize WinMD generator to match other build task conventions
Align WinRT.WinMD.Generator with the patterns used by WinRT.Interop.Generator,
WinRT.Impl.Generator, and WinRT.Projection.Generator:

csproj:
- Add IsBuildTool with conditional RuntimeIdentifier/SelfContained for publishing
- Add common metadata properties (Description, AssemblyTitle, Copyright, etc.)
- Split NativeAOT settings into separate PropertyGroup with missing settings:
  ControlFlowGuard, IlcDehydrate, IlcResilient, UseSystemResourceKeys (Release)
- Change LangVersion from 14.0 to preview (matches Impl/Projection)

Error handling:
- Add WellKnownWinMDException base class with error ID and formatted ToString()
- Add UnhandledWinMDException with phase tracking and bug-report message
- Update WellKnownWinMDExceptions to use error code prefix CSWINRTWINMDGEN
  with numbered IDs (0001-0006) instead of generic InvalidOperationException
- Add Extensions/WinMDExceptionExtensions.cs with IsWellKnown extension

Generator structure:
- Make WinMDGenerator a partial class split by phase:
  - WinMDGenerator.cs: Run() with try-catch-when per phase
  - WinMDGenerator.Discover.cs: assembly loading and type discovery
  - WinMDGenerator.Generate.cs: WinMD generation and writing
- Add WinMDGeneratorDiscoveryState to pass state between phases

WinmdWriter decomposition:
- Split 1819-line WinmdWriter.cs into partial files:
  - WinmdWriter.cs: core (fields, ProcessType, FinalizeGeneration, Write)
  - WinmdWriter.Types.cs: enum/delegate/interface/struct/class processing
  - WinmdWriter.SynthesizedInterfaces.cs: synthesized + custom mapped interfaces
  - WinmdWriter.Members.cs: method/property/event helpers
  - WinmdWriter.TypeMapping.cs: type signature mapping + type references
  - WinmdWriter.Attributes.cs: attribute creation, copying, and helpers
- Move TypeDeclaration class to Models/TypeDeclaration.cs
- Clean up unnecessary using directives per file
