+++ src/WinRT.WinMD.Generator/Discovery/AssemblyAnalyzer.cs
