+++ src/WinRT.WinMD.Generator/Extensions/WinMDExceptionExtensions.cs
@@ -1,0 +1,21 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System;
+using WindowsRuntime.WinMDGenerator.Errors;
+
+namespace WindowsRuntime.WinMDGenerator;
+
+/// <summary>
+/// Extensions for WinMD generator exceptions.
+/// </summary>
+internal static class WinMDExceptionExtensions
+{
+    extension(Exception exception)
+    {
+        /// <summary>
+        /// Gets a value indicating whether an exception is well known (and should therefore not be caught).
+        /// </summary>
+        public bool IsWellKnown => exception is OperationCanceledException or WellKnownWinMDException;
+    }
+}
