+++ src/WinRT.WinMD.Generator/Generation/WinMDGeneratorArgs.Parsing.cs
