+++ src/WinRT.WinMD.Generator/Generation/WinMDGeneratorArgs.cs
