+++ src/WinRT.WinMD.Generator/Models/TypeMapper.cs
