+++ src/WinRT.WinMD.Generator/Program.cs
