+++ src/WinRT.WinMD.Generator/Resolvers/PathAssemblyResolver.cs
