+++ src/WinRT.WinMD.Generator/WinRT.WinMD.Generator.csproj
@@ -2,25 +2,29 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net10.0</TargetFramework>
-    <LangVersion>14.0</LangVersion>
+    <LangVersion>preview</LangVersion>
     <Nullable>enable</Nullable>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <PublishAot>true</PublishAot>
     <DisableRuntimeMarshalling>true</DisableRuntimeMarshalling>
-    <InvariantGlobalization>true</InvariantGlobalization>
-    <OptimizationPreference>Speed</OptimizationPreference>
-    <StackTraceSupport>false</StackTraceSupport>
+
+    <!-- Common metadata properties for publishing -->
+    <Description>C#/WinRT WinMD Generator v$(VersionString)</Description>
+    <AssemblyTitle>C#/WinRT WinMD Generator v$(VersionString)</AssemblyTitle>
+    <Copyright>Copyright (c) Microsoft Corporation. All rights reserved.</Copyright>
+    <AssemblyVersion>$(AssemblyVersionNumber)</AssemblyVersion>
+    <FileVersion>$(VersionNumber)</FileVersion>
 
     <!-- Emit the '[module: SkipLocalsInit]' attribute (same as the runtime) -->
     <EmitSkipLocalsInitAttribute>true</EmitSkipLocalsInitAttribute>
       
-    <!-- Set the root namespace -->
+    <!-- Set the root namespace to match the other projects -->
     <RootNamespace>WindowsRuntime.WinMDGenerator</RootNamespace>
 
     <!-- This project produces a CLI tool, so we want a nice short name for it -->
     <AssemblyName>cswinrtwinmdgen</AssemblyName>
       
-    <!-- Same settings as WinRT.Runtime2 -->
+    <!-- Same settings as 'WinRT.Runtime' -->
     <TreatWarningsAsErrors Condition="'$(Configuration)' == 'Release'">true</TreatWarningsAsErrors>
     <CodeAnalysisTreatWarningsAsErrors Condition="'$(Configuration)' == 'Release'">true</CodeAnalysisTreatWarningsAsErrors>
     <AnalysisLevel>latest</AnalysisLevel>
@@ -28,6 +32,22 @@
     <EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
     <Features>strict</Features>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
+
+    <!-- Used to enable publishing project in build -->
+    <IsBuildTool>true</IsBuildTool>
+    <RuntimeIdentifier Condition="'$(PublishBuildTool)' == 'true'">win-$(BuildToolArch)</RuntimeIdentifier>
+    <SelfContained Condition="'$(PublishBuildTool)' == 'true'">true</SelfContained>
+  </PropertyGroup>
+
+  <!-- NativeAOT and optimization options (same as 'WinRT.Interop.Generator') -->
+  <PropertyGroup>
+    <InvariantGlobalization>true</InvariantGlobalization>
+    <OptimizationPreference>Speed</OptimizationPreference>
+    <StackTraceSupport Condition="'$(Configuration)' == 'Release'">false</StackTraceSupport>
+    <UseSystemResourceKeys Condition="'$(Configuration)' == 'Release'">true</UseSystemResourceKeys>
+    <ControlFlowGuard>Guard</ControlFlowGuard>
+    <IlcDehydrate>false</IlcDehydrate>
+    <IlcResilient>false</IlcResilient>
   </PropertyGroup>
 
   <ItemGroup>
